+++ src/main/java/app/models/Bidder.java
@@ -18,7 +18,7 @@ public class Bidder extends User implements AuctionObserver {
   }
 
   @Override
-  public void onNewBidPlaced(String itemName, double newPrice, String bidderName) {
+  public void onNewBidPlaced(String itemName, long newPrice, String bidderName) {
     if (this.getName().equals(bidderName)) {
       System.out.println("[" + getName() + "]: Bạn đã giữ giá cao nhất mới: $" + newPrice);
     } else {
@@ -35,7 +35,7 @@ public class Bidder extends User implements AuctionObserver {
   }
 
   @Override
-  public void onAuctionClosed(String itemName, String winnerName, double finalPrice) {
+  public void onAuctionClosed(String itemName, String winnerName, long finalPrice) {
     if (this.getName().equals(winnerName)) {
       System.out.println(
           "["
