+++ src/main/java/app/models/Seller.java
@@ -18,7 +18,7 @@ public class Seller extends User implements AuctionObserver {
   }
 
   @Override
-  public void onNewBidPlaced(String itemName, double newPrice, String bidderName) {
+  public void onNewBidPlaced(String itemName, long newPrice, String bidderName) {
     System.out.println(
         "["
             + getName()
@@ -31,7 +31,7 @@ public class Seller extends User implements AuctionObserver {
   }
 
   @Override
-  public void onAuctionClosed(String itemName, String winnerName, double finalPrice) {
+  public void onAuctionClosed(String itemName, String winnerName, long finalPrice) {
     System.out.println(
         "["
             + getName()
