+++ src/main/java/app/obserser/AuctionObserver.java
@@ -1,7 +1,7 @@
 package app.obserser;
 
 public interface AuctionObserver {
-  void onNewBidPlaced(String itemName, double newPrice, String bidderName);
+  void onNewBidPlaced(String itemName, long newPrice, String bidderName);
 
-  void onAuctionClosed(String itemName, String winnerName, double finalPrice);
+  void onAuctionClosed(String itemName, String winnerName, long finalPrice);
 }
