+++ src/main/java/app/obserser/Subject.java
@@ -5,5 +5,5 @@ public interface Subject {
 
   void removeObserver(Observer observer);
 
-  void notifyObserversNewBid(double price, String bidderName);
+  void notifyObserversNewBid(long price, String bidderName);
 }
